requirements.txt
requests==2.27.1
folium==0.12.1.post1
geopy==2.2.0
Flask==2.1.2
python-dotenv==0.20.0

